+++ src/main/java/org/projectschool/school/core/controller/PersonController.java
@@ -2,7 +2,9 @@ package org.projectschool.school.core.controller;
 
 import org.projectschool.school.core.bs.dao.PersonRepository;
 import org.projectschool.school.core.eis.bo.Person;
+import org.projectschool.school.core.util.BaseResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,22 +22,50 @@ public class PersonController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public Object getOnePerson(@PathVariable("id") Long id){
-        return personRepository.findOne(id);
+        if(id == null || id <= 0){
+            return new BaseResponse(HttpStatus.CONFLICT.value(), HttpStatus.CONFLICT, "Debe mandar un id válido. ¡El id: " + id + " no se reconoce en la DB.!");
+        }
+
+        if(personRepository.findOne(id) == null){
+            return new BaseResponse(HttpStatus.NOT_FOUND.value(), HttpStatus.NOT_FOUND,"No se encuentra ese Person en la DB.");
+        }
+
+        return new BaseResponse(HttpStatus.FOUND.value(), HttpStatus.FOUND, "Operación realizada correctamente.", personRepository.findOne(id));
     }
 
     @RequestMapping(method = RequestMethod.POST)
     public Object savePerson(@RequestBody(required = true) Person person){
-        return this.personRepository.save(person);
+        return new BaseResponse(HttpStatus.ACCEPTED.value(), HttpStatus.ACCEPTED, "Person registrado correctamente.", this.personRepository.save(person));
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PATCH)
     public Object editPerson(@PathVariable("id") Long id, @RequestBody(required = true) Person newPerson){
-        return personRepository.save(newPerson);
+
+        if(id == null || id <= 0){
+            return new BaseResponse(HttpStatus.CONFLICT.value(), HttpStatus.CONFLICT, "Debe mandar un id válido. ¡El id: " + id + " no se reconoce en la DB.!");
+        }
+
+        if(personRepository.findOne(id) == null){
+            return new BaseResponse(HttpStatus.NOT_FOUND.value(), HttpStatus.NOT_FOUND,"No se encuentra ese Person en la DB.");
+        }
+
+        return new BaseResponse(HttpStatus.ACCEPTED.value(), HttpStatus.ACCEPTED, "Person editado correctamente.", personRepository.save(newPerson));
+
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public void deletePerson(@PathVariable("id") Long id){
+    public Object deletePerson(@PathVariable("id") Long id){
+        if(id == null || id <= 0){
+            return new BaseResponse(HttpStatus.CONFLICT.value(), HttpStatus.CONFLICT, "Debe mandar un id válido. ¡El id: " + id + " no se reconoce en la DB.!");
+        }
+
+        if(personRepository.findOne(id) == null){
+            return new BaseResponse(HttpStatus.NOT_FOUND.value(), HttpStatus.NOT_FOUND,"No se encuentra ese person en la DB.");
+        }
+
         personRepository.delete(id);
+
+        return new BaseResponse(HttpStatus.ACCEPTED.value(), HttpStatus.ACCEPTED, "Person eliminado correctamente.");
     }
 
     /* Format base JSOn
